+++ usuarios/views/__init__.py
@@ -1,0 +1,6 @@
+from .usuarios import (
+    lista_usuarios,
+    novo_usuario,
+    editar_usuario,
+    inativar_usuario,
+)